Remove raspi platform dep on crashpad wrapper stub (#7649)
This dependency causes the loader app to have dependency paths on both
the real and stub crashpad wrapper libraries, and we observed the stub
winning during linking. Starboard platforms no longer need to depend on
the crashpad wrapper (real or stub), as this integration is now handled
by the loader_app and elf_loader_sandbox.

Issue: 453749661