[Android] Flush cookies and local storage when activity is paused (#7587)
Similar to c25 which flushes cookies and localStorage on the cobalt
freeze event, this PR adds flushing cookies and localStorage to the
Android activity onPause() handler.

In addition to this, flush to cookies and localStorage is performed when
h5vcc.system.exit() is called in support of YTS tests. In production the
flush may be redundant but will act as a no-op if it is.

Issue: 371433332