Replace std::bind_front with lambdas for cxx17 (#7749)
Customizations made as part of the m138 rebase to remove the use of
std::bind_front - added in cxx20 - for cxx17 builds fixed the compiler
error but introduced an undefined symbol. This PR restores the original
function signatures to resolve the undefined symbol and uses lambdas,
instead of std::bind_front, to achieve partial function application for
cxx17 builds.

Issue: 454950085