android: parallelize web contents loading during app startup (#7713)
Introduce a parallel loading mechanism for web content during app
startup.
A dedicated 'app' shell is now created to load the main application URL
in the background. Concurrently, a 'splash' shell can display initial
content or a placeholder. Once the main app finishes loading, the app
shell replaces the splash shell.

This change aims to improve perceived startup performance on Android
devices by overlapping the loading of the main application with the
display of an initial splash screen. New command-line arguments and
AndroidManifest metadata are added to configure the splash URL, timeout,
and to optionally disable the native splash experience.

This PR is mostly the same as
https://github.com/youtube/cobalt/pull/7456.

Issue: 441738743