Cherry pick PR #7131: evergreen: more customizations to update_client (#7738)
This is a manual cherry-pick with no sustantial changes. Some legacy
functions and classes in m114 are ported over because the Cobalt
updater customizations depend on them. TODOs are filed and included in
the files. Follow-up PRs will resolve them, and this PR is kept as a
cherry-pick.

TODO(454962891): Investigate the upstream method to cancel a download
TODO(455004672): Investigate the legacy classes and functions from m114.

Original PR #7131 
These customizations support in-memory update and installation slot
manipulation. They mostly come from C25, with modifications to work with
the new codebase. Added a new CrxDownloaderFactoryCobalt class.

Issue: 431862767