Include starboard player test data in the test bundles (#7794)
These test files are required for e.g. nplb to run; if they are not
found we hit a CHECK early on and cannot even list the available tests.

To avoid requiring people to know that they need to depend on this new
target, make it the only possible dependency by defining it for all
platforms (on non-Apple ones it is just a group with `data_deps` set to
"player_download_test_data") and making "player_download_test_data" an
internal target that other files do not depend on explicitly.

Bug: 433180638