[tvOS] Remove posix_emu/ directory
This directory includes a `pthread_setname_np()` implementation with a
signature that matches glibc's and musls's (by taking 2 arguments instead of
1), and it came from C25.

Although it works as expected with Starboard code (once plugged into the
build system), it breaks when called from Chromium code. Chromium code
already calls `pthread_setname_np()` with the right number of arguments,
which leads to a wrong comparison that always fails in this
code (pthread_self() being compared to a `const char*`) and a lot of
misleading error messages about failing to set the thread name being printed
on the console.

Since most usages of `pthread_setname_np()` have been fixed to work with
Apple's version in PR #7052 and the remaining case is being updated in this
commit, we can drop this workaround altogether.

(posix_emu also has a memrchr.c that is being removed here. It came from C25
and is not used anywhere at the moment)

Bug: 432507888
Bug: 456692793