Enable optimize_for_size for 3P
Turn on the optimize_for_size wich for the most
part just enables -Os instead of -O2. The flag is on
for Android but for some reason was not turned of for 3P.
This optimization was also used in Cobalt 25.

The result is 14MB binary size reduction and ~ 10MB memory reduction.

 libcobalt.so 85MB -> 71MB
 libcobalt.lz4 52MB -> 43MB
 RSS home, signed out 135MB -> 124MB

Fix the linux-x64x11 linking. Dead code was not removed:
  mojo/core/node_channel.cc

Fix evergreen-x64 builds. The compiler generate references to:
  __ashlti3
  __lshrti3

Fix arm builds. The copiler generate references to:
  __aeabi_lasr
  __aeabi_llsl
  __aeabi_llsr

Issue: 421171908