cobalt/test: Create new browser test data target (#7845)
This change adds a data target that composes Cobalt browser test's data
dependencies. This was instrumented by a script that basically links to
all of the resource files underneath `//content/test/data`. The script
filters out only the symlink files. The script and its unittest script
were added to the data directory for updating the data directory for
each build, which should enable us to maintain this dependency more
easily after future rebases.

Cobalt tests failed recently when trying to use symlinks in tests. On
the test devices, this was disallowed because they used FAT file
systems, where symlinks are not compatible. This will require more
granular updates to the test data target, but this should be fine and
should be a relatively low-effort extra step.

This also removes a file chooser test, which does not reflect a Cobalt
CUJ.

Test: `python3 -m unittest
cobalt/testing/browser_tests/data/update_build_gn_test.py`
Test: `out/android-x86_devel/bin/run_cobalt_browsertests -v  -f`
Bug: 432313695