[media m138] fix side data issue (#8001)
DecoderBufferSideData was added as a formalized data structure in:
https://chromium-review.googlesource.com/c/chromium/src/+/4757645 

Since https://chromium-review.googlesource.com/c/chromium/src/+/5918375
an empty DecoderBufferSideData is also used to in an EOS to tell the
decoder that it should flush its current state and prepare for a new
configuration.

This PR adds an additional check to ensure the alpha_data vector has
contents before copying.

Tested on Boreal, can trigger the DCHECK without this change at the end
of an ad, and passes with the change.

Bug: 456780532