android: Independent timeout for network error dialog (#8003)
The network error dialog is seeing some edge cases where it is not
displaying (ie. failed DNS resolution) and thus causing a black screen.
This is due to the fact the the `urlConnection` timeouts only begin once
DNS lookup is finished. This PR adds an independent timeoutHandler that
will display the error dialog if the check is not successful within 10
seconds, regardless of a successful/non-successful/hanging DNS
connection.

Bug: 394176365
Bug: 457471731

---------

Co-authored-by: johnx <johnx@google.com>