media: Expose decoder video budget clamp to h5vcc (#7964)
This PR adds the ability for us to experiment with reducing the size of
the video DecoderBuffer pool. This is done through a new
`memory_limit_clamp_` value in `SourceBufferStream`, which will clamp
the value of `memory_limit_` if `memory_limit_` exceeds
`memory_limit_clamp_`. The `memory_limit_clamp_` value is then exposed
to an h5vcc setting, allowing us to set custom values for this clamp.

To get this implemented two major things were done:

1. In `cobalt_content_renderer_client.cc`, we query for the h5vcc value
during `GetStarboardRendererFactoryTraits`. If it exists, we enable a
new media switch `kMSEVideoBufferSizeLimitClampMb`, associating this
switch with the set value from h5vcc.

2. When configuring a new video config, we set its memory limit to the
smaller value between `SbMediaGetVideoBufferBudget` or our set clamp
value, which is retrieved from the media switch.

Bug: 416105795
Bug: 460292554