[tvOS] BACKPORT: Prevents forwarding handled events to super.
Backport @jkim-julie's https://crrev.com/c/7149563 from M144. This is
another CL in the series of required backports to properly handle input from
the remote control on tvOS.

Bug: 456400530
Bug: 456692793
Bug: 432824509

---

Original CL description:

This prevents forwarding handled events to super. On tvOS 26, keyboard
input can trigger duplicate remote button events, so handled events
are not propagated. UIPressTypeMenu events are an exception and are
still passed to super to allow to move the app to the background.

Change-Id: Iedcf58b463e09a124232f2f5c167fb1fd1c80461
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7149563
Cr-Commit-Position: refs/heads/main@{#1545408}