Fix DNS lookups for IPv6 addresses. (#8053)
The Starboard layer wrapper for getaddrinfo() made incorrect copies of
the address returned for IPv6 addresses.

This fixes that and adds extensive testing using a host with known IP
addresses.

The fix is adopted from https://github.com/youtube/cobalt/pull/7588 with
a simplification and additional tests.

This PR also includes an update to test certs for net_unittests that
expired Oct 29 2025.

Bug: 460175900
Bug: 438858995
Bug: 409716671
Bug: 460594994
Bug: 439002238
Bug: 405988446

---------

Co-authored-by: jimmy.ge <jimmy.ge@vidaa.com>