Migrate runner's build directory to /runner-cache (#8000)
This is shadowing PR following the sandbox [PR
815](https://github.com/youtube/cobalt_sandbox/pull/815).
This change allows runners to mount /runner-cache from the host, thus
the tool chain files and dependent repo fetched by gclient could be used
for the next runner.

After this change, below files will be cached crossing runners on the
same node:

- checked in code on cobalt/src: to save the artifacts downloaded by
gclient runhook, mainly the tool chain files.
- CIPD cache
- source code fetched by gclient sync
As a results, the main checkout will move to ~/cobalt/src(a soft link to
/runner-cache/cobalt) from ~/src.

Bug: 407051026