android: refactor activeNetworkCheck() to use ScheduledThreadPool (#8023)
This follow-up CL to #8003 addresses closing threads that might have
been left hanging with the newly implemented timeout. It refactors
`activeNetworkCheck()` to use `ScheduledThreadPool` instead of the
alternative of trying to interrupt or cancel them manually, which may or
may not work due to the implementation of `HttpURLConnection`. It also
sets a few methods on the URL request to make it more thorough and
closer to Chromium's version.

Bug: 394176365
Bug: 457471731
(cherry picked from commit cacd413b3ac41a631371441c1407829e5fb91921)