Restrict optimizations to Starboard + Android, add a flag to control it
Add the ENABLE_MULTI_PACKETS_PER_CALL_QUIC_OPTIMIZATIONS build flag to
control whether the code enabling these optimizations will be built, and set
it to true when building for either Starboard or Android.

The rationale for restricting to these two platforms is that in both cases
the kernel being run is the Linux kernel, which these optimizations depend
on.

The base declaration/implementation in net/socket/socket.{cc,h} remains
restricted only to IS_COBALT, as it just returns ERR_NOT_IMPLEMENTED by
default and allows the rest of the code to handle the failure without
requiring more `#ifdef`s.