Cherry pick PR #6470: media: Add BASE feature for Android Overlay (#8082)
Refer to the original PR: https://github.com/youtube/cobalt/pull/6470

Implements a BASE Feature, `CobaltUsingAndroidOverlay`, for SbPlayer
using `AndroidOverlay`, and this is disabled by default.

This feature is used to enable `StarboardRenderer` to create
`AndroidOverlay`, so SbPlayer can use the SurfaceView of
`AndroidOverlay`, instead of VideoSurfaceView from StarboardBridge.

This ensures the lifecycle of SurfaceView is managed by
StarboardRenderer, instead of in StarboardBridge.

Issue: 428017363
Issue: 460831614