[tvOS] Fix obj-c string checking logic (#8040)
The previous logic was checking for memory address equality and doing a
memory assignment, which will cause the check to always be true. In
objc, need to use `isEqualToString:` to correctly compare NSString and
use [... copy] to assign values.

Checked all other starboard/tvos files and found this was the only
usage.

Bug: 460583812