media: Reshape AndroidOverlayFactoryCB ownership (#8075)
Move the AndroidOverlayMojoFactoryCB from GpuMojoMediaClient to
GpuMojoMediaClientStarboard. This localizes the Android-specific
overlay factory management to the Starboard implementation, where it
is exclusively used.

Additionally, internalize the CreatePlatformStarboardRenderer free
function into GpuMojoMediaClientStarboard as a virtual method,
further improving modularity and separation of concerns within the
media pipeline for platform-specific rendering. This revisits and
refines earlier changes to the media client architecture.

This revisits the changes in
https://github.com/youtube/cobalt/commit/7c7e771e9a42a2bd9585efd562589a8ad5430563.

Issue: 433534482
Issue: 429257581