starboard/linux: Refactor NetworkNotifier to use starboard::Thread
Refactor NetworkNotifier in starboard/shared/linux/system_network_status.cc
to use starboard::Thread instead of raw pthread.

This change also switches NetworkNotifier to use a joinable thread. Since
NetworkNotifier is a singleton that lives for the lifetime of the process,
using a joinable thread is safe (the OS cleans it up) and simplifies the
implementation by avoiding detached thread creation.

Additionally:
- Updates is_online_ to be atomic to prevent data races.
- Adds error checking for socket creation and fixes a resource leak by
  closing the socket on thread exit.