tvOS: Connect menu press events in Cobalt Shell and Starboard
Patch by @jkim-julie. It is the last in the series of changes required to
support the C25 flow of not immediately switching to the home screen
when the Menu button is pressed on the remote.

Instead, pressing the Menu button acts like a "back" button that goes back
different screens in the YT app until it reaches a point in which
H5vccSystemImpl::PerformExitStrategy() is called and the app is suspended.

More specifically:

1. Implement `pressesBegan()` and `pressesEnded()` in
   shell_platform_delegate_ios.mm. They call registerMenu{PressBegan,
   PressEnded}.
2. Implement registerMenu{PressBegan, PressEnded} storing the events at
   SDBStarboardApplication (i.e. `application_darwin.mm`).
3. When `H5vccSystemImpl::PerformExitStrategy()` is called, it just invokes
   `SbSystemRequestConceal()`, which is just a wrapper for calling
   `SBDStarboardApplication::suspendApplication()`.
4. Make `SBDStarboardApplication::suspendApplication()` send those
   events to the UIApplication (i.e. not one of its child UIViews) so that
   it is processed as expected.

Bug: 447135715
Bug: 456400530