Always flush cookies on exit (#8133)
This is an attempt to increase cookie storage robustness.

Related to #8131, this enforces flushing cookies on app exit every time.

Note that there's a small risk in blocking and delaying the exit here
when a synchronous disk flush takes place. However, partition flush
translates to `fdatasync()` which typically takes milliseconds not to
`fsync()` which can lock up the system for up to a second. ( see #3578 )

Bug: 461896983
Bug: 371433332
Bug: 451215656
Bug: 450936106
Bug: 456836622
Bug: 444451968
Bug: 448130250