tvOS: Remove the custom EGL code from the tree
This code is not working as expected in Chrobalt, and is mostly used for
decode to texture, which is not working in main even for other platforms.

In the main branch, we rely on ANGLE directly, and having this code around
prevents us from removing other unused functions, variables and files. With
this code gone, it does not make sense to keep its consumers (e.g.
video_decoder.mm and vpx_video_decoder.mm) in the tree. They were used only
in decode to texture mode, so they were already broken anyway.

Bug: 447334535
Bug: 459507205