feat: Use android-arm tests for android-x86 coverage and fix pylint issues
This change modifies the behavior of the coverage baseline script.
When running coverage for the android-x86 platform, the script will
now use the unittests defined for the android-arm platform. This is
necessary because the android-x86 platform does not have its own
set of unittests. Also, fixed all pylint issues and import errors in
both establish_coverage_baseline.py and establish_coverage_baseline_test.py.