media: Add AllocateOnDemand experiment (#8119)
This PR exposes the `allocate_on_demand` setting for
DecoderBufferAllocator to H5VCC experiments. This allows us to
explicitly enable `allocate_on_demand` at runtime.

To make this happen, the following changes have been made:

- |is_memory_pool_allocated_on_demand_| is no longer const, as this
experiment enables the possibility of it being changed while
`DecoderBufferAllocator` exists.
- Locking the mutexes for `Suspend()` and `Resume()` earlier in the
function call, as it is possible that
|is_memory_pool_allocated_on_demand_| is changing when these functions
are called.
- `cobalt_content_renderer_client.cc` will now query for this specific
setting, and will update the value as necessary.

Bug: 459199176