tvos: AVSBVideoRenderer: fetch a valid AVDisplayManager instance for use
`ApplicationWindow` was removed in #8366, but even before that this code did
not work as expected in main because in main `SbWindowCreate()` was not
invoked and therefore the attempt to get the `avDisplayManager` property
just did nothing.

Make the code work by fetching a valid `UIWindow`, the one created in
`shell_platform_delegate_ios.mm`, instead. And make sure to link against the
AVKit framework, otherwise accessing the `avDisplayManager` property causes
a "invalid selector" crash at runtime.

Bug: 456692793
Bug: 459366150