[v8] BACKPORT: Move v8_enable_drumbrake to v8.gni (#8196)
Backport https://crrev.com/c/6768061 by @jkim-julie. This is part of the
work to get interpreted WASM to work for tvOS.

Bug: 462369287

---

Original CL description:

This moves v8_enable_drumbrake to v8.gni from BUILD.gn. Since it's
related to WebAssembly, it would be good to be placed at the same file
with v8_enable_webassembly. It's a precursor CL to configure build
options for tvOS that needs v8_enable_drumbrake.

Change-Id: I0af3bbb579a3a84b5cd233af30cee1fea449c5b8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6768061
Cr-Commit-Position: refs/heads/main@{#101576}

Co-authored-by: Julie Jeongeun Kim <jkim@igalia.com>