[tvOS] Make tvOS code in Starboard use ANGLE (#8223)
Set `gl_type="angle"` just like the linux-x64x11 port already does,
which causes the `//starboard/egl_and_gles` GN config properly add the
ANGLE dependencies to the `starboard_platform` target.

This also requires teaching FakeGraphicsContextProvider to use ANGLE's
Metal backend on Apple platforms, as the OpenGL ES backend is not
available at least on tvOS.

Note that this makes the custom EGL/GLES code in the tree, which came
from C25, even more broken. It is only used for decode to texture (which
is broken for all ports), and will be removed in a follow-up PR.

Bug: 456692793
Bug: 447334535