Enable in-memory update (#8161)
Download the update package to a large string in memory, instead of a
file, in order to save required storage.

The vast majority of the code is ported from the original PR #849 to
enable this feature, with the following modifications:
1. NetworkFetcher::DownloadToString() utilizes SimpleURLLoader instead
of URLFetcher, to be consistent with the upstream change.
2. Increases SimpleURLLoader::kMaxBoundedStringDownloadSize to 100MB for
Cobalt, because as tested locally, the QA package size is 70MB. The size
range of the official release packages is unknown yet, because the infra
is WIP. This number is open to change.
3. Relatively minor change in zip::Unzip(), to work with the updated
upstream code.

Bug: 444006168
Bug: 431862767